Handle buttons and LEDs server-side
Stateless logic in console. Remove knowledge of current status in the
console. The console, once powered, sends periodically a keep-alive
message to the remote server. Its payload can contain also indication
that a button has been pressed, with longer than 5 seconds pressures
being treated as a different action (NEW!).
The remote server replies with a status message telling the console how
to light up its three LEDs.

This change should allow the remote server to detect when the console is
powered up and when it gets disconnected or powered off. There will be
no need for a start and stop button on the console.
The 6 actions related to the 3 buttons can be new configured entirely
server-side without the need to reprogram the console module.